+++ src/app/list/ShoppingRecipeItem.tsx
@@ -3,6 +3,7 @@
 import { Button } from "~/components/ui/button";
 import { useShoppingListActions } from "../useShoppingListActions";
 import { Trash } from "lucide-react";
+import { SimpleAlertDialog } from "~/components/SimpleAlertDialog";
 
 export function ShoppingRecipeItem(props: { id: string; name: string }) {
   const { id, name } = props;
@@ -11,15 +12,22 @@ export function ShoppingRecipeItem(props: { id: string; name: string }) {
 
   return (
     <div key={id} className="flex items-center gap-2">
-      <Button
-        onClick={async () => {
+      <SimpleAlertDialog
+        trigger={
+          <Button variant="destructive-outline" size="sm">
+            <Trash />
+          </Button>
+        }
+        title={"Are you sure you want to delete?"}
+        description={
+          "This will remove all items for this recipe from your shopping list."
+        }
+        confirmText={"Delete"}
+        cancelText={"Cancel"}
+        onConfirm={async () => {
           await handleDeleteRecipe(Number(id));
         }}
-        variant="destructive-outline"
-        size="sm"
-      >
-        <Trash />
-      </Button>
+      />
       <span className="break-words text-lg font-semibold">{name}</span>
     </div>
   );
